menu refactor

+++ applet/src/menu.c
@@ -692,7 +692,7 @@ void create_menu_entry_edit_screen(void)
     uint8_t *p;
 
     md380_menu_0x2001d3c1 = md380_menu_0x200011e4;
-    mn_editbuffer_poi = (uint32_t) md380_menu_edit_buf;
+    mn_editbuffer_poi = md380_menu_edit_buf;
 
 
     /*
@@ -713,6 +713,7 @@ void create_menu_entry_edit_screen(void)
      */
 
     // clear retrun buffer //  see 0x08012a98
+    // TODO: is wchar_t (16 bits))
     for (i = 0; i < 0x11; i++) {
         p = (uint8_t *) mn_editbuffer_poi;
         p = p + i;
@@ -798,11 +799,12 @@ void create_menu_entry_edit_dmr_id_screen(void)
     uint32_t nchars;
 
     md380_menu_0x2001d3c1 = md380_menu_0x200011e4;
-    mn_editbuffer_poi = (uint32_t) md380_menu_edit_buf;
+    mn_editbuffer_poi = md380_menu_edit_buf;
 
 
 
     // clear retrun buffer //  see 0x08012a98
+    // TODO: is wchar_t (16 bits))
     for (i = 0; i < 0x11; i++) {
         p = (uint8_t *) mn_editbuffer_poi;
         p = p + i;
